+++ .gitignore
@@ -1,5 +1,5 @@
 node_modules
-HELP.md
+backend/HELP.md
 target/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
