+++ backend/src/main/java/com/example/sematewebshop/applikation/CatalogService.java
