+++ backend/src/main/java/com/example/sematewebshop/applikation/CustomerService.java
@@ -1,4 +1,5 @@
 package com.example.sematewebshop.applikation;
 
 public class CustomerService {
+
 }
