+++ backend/src/main/java/com/example/sematewebshop/docu.txt
