+++ backend/src/main/java/com/example/sematewebshop/domain/Customer.java
