+++ backend/src/main/java/com/example/sematewebshop/domain/Product.java
