+++ backend/src/main/java/com/example/sematewebshop/model/ProductDetailDTO.java
