+++ backend/src/main/java/com/example/sematewebshop/model/ProductOverviewDTO.java
