+++ backend/src/main/java/com/example/sematewebshop/persistenz/JpaUserRepository.java
@@ -1,0 +1,10 @@
+package com.example.sematewebshop.persistenz;
+
+import com.example.sematewebshop.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface JpaUserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+}
