+++ backend/src/main/resources/application.properties
@@ -1,2 +1,1 @@
-vaadin.launch-browser=true
 spring.application.name=SeMateWebshop
