--- src/main/java/com/example/sematewebshop/domain/RegisterRequest.java
@@ -1,15 +1,0 @@
-package com.example.sematewebshop.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-public class RegisterRequest {
-    @Getter
-    @Setter
-    private String username;
-    @Getter
-    @Setter
-    private String password;
-
-}
-
